fix(studio): cleanup the PR, moving, renaming files, remove tests
