--- web/packages/studio/src/components/sidePanels/AgentPanels/AgentPanel/types.ts
@@ -1,4 +1,0 @@
-// SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-export type AgentPanelTab = 'agent-details' | 'chat-playground' | 'deployment-logs';
