+++ web/packages/studio/src/routes/agents/AgentEvaluationsRoute/AgentEvaluationsListRoute.tsx
@@ -4,9 +4,9 @@
 import { Button, PageHeader, Stack } from '@nvidia/foundations-react-core';
 import { AccessibleTitle } from '@studio/components/AccessibleTitle';
 import { AgentEvaluationsDataView } from '@studio/components/dataViews/AgentEvaluationsDataView';
+import { SubmitEvaluationModal } from '@studio/components/evaluation/SubmitEvaluationModal';
 import { useWorkspaceFromPath } from '@studio/hooks/useWorkspaceFromPath';
 import { useBreadcrumbs } from '@studio/providers/breadcrumbs/useBreadcrumbs';
-import { SubmitEvaluationModal } from '@studio/routes/agents/AgentEvaluationsRoute/components/SubmitEvaluationModal';
 import { getAgentsListRoute } from '@studio/routes/utils';
 import { useState, type FC } from 'react';
